PR comments rename train -> build_trained
