+++ phygnn/model_interfaces/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 """Model Interfaces"""
 from .phygnn_model import PhygnnModel
+from .random_forest_model import RandomForestModel
 from .tf_model import TfModel
