+++ phygnn/utilities/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 """Utilities"""
 from .pre_processing import PreProcess
-from .tf_utilities import tf_isin, tf_log10, Layers
+from .tf_layers import Layers
+from .tf_utilities import tf_isin, tf_log10
