+++ tests/test_phygnn_model.py
@@ -67,11 +67,11 @@ def p_fun_pythag(y_predicted, y_true, p):
 def test_nn():
     """Test the basic NN operation of the PGNN without weighting pfun."""
     PhysicsGuidedNeuralNetwork.seed(0)
-    model = PhygnnModel.train(p_fun_pythag, features, labels, P,
-                              hidden_layers=HIDDEN_LAYERS,
-                              loss_weights=(1.0, 0.0),
-                              n_batch=4,
-                              n_epoch=20)
+    model = PhygnnModel.build_trained(p_fun_pythag, features, labels, P,
+                                      hidden_layers=HIDDEN_LAYERS,
+                                      loss_weights=(1.0, 0.0),
+                                      n_batch=4,
+                                      n_epoch=20)
 
     test_mae = np.mean(np.abs(model.predict(X, table=False) - Y))
 
@@ -86,11 +86,11 @@ def test_nn():
 def test_phygnn_model():
     """Test the operation of the PGNN with weighting pfun."""
     PhysicsGuidedNeuralNetwork.seed(0)
-    model = PhygnnModel.train(p_fun_pythag, features, labels, P,
-                              hidden_layers=HIDDEN_LAYERS,
-                              loss_weights=(0.0, 1.0),
-                              n_batch=4,
-                              n_epoch=20)
+    model = PhygnnModel.build_trained(p_fun_pythag, features, labels, P,
+                                      hidden_layers=HIDDEN_LAYERS,
+                                      loss_weights=(0.0, 1.0),
+                                      n_batch=4,
+                                      n_epoch=20)
 
     test_mae = np.mean(np.abs(model.predict(X, table=False) - Y))
 
@@ -111,12 +111,12 @@ def test_phygnn_model():
 def test_normalize():
     """Test the operation of the PGNN with weighting pfun."""
     PhysicsGuidedNeuralNetwork.seed(0)
-    model = PhygnnModel.train(p_fun_pythag, features, labels, P,
-                              normalize=False,
-                              hidden_layers=HIDDEN_LAYERS,
-                              loss_weights=(0.0, 1.0),
-                              n_batch=4,
-                              n_epoch=20)
+    model = PhygnnModel.build_trained(p_fun_pythag, features, labels, P,
+                                      normalize=False,
+                                      hidden_layers=HIDDEN_LAYERS,
+                                      loss_weights=(0.0, 1.0),
+                                      n_batch=4,
+                                      n_epoch=20)
 
     test_mae = np.mean(np.abs(model.predict(X, table=False) - Y))
